requirements.txt
kivy
kivymd 
pyotp 
qrcode
pillow 
pyzbar
opencv-python